New MIDI_Event selection scheme - state transition as key, ...
I realized that, probably, what is needed to select the right MIDI_Event
subtype (to dispatch the last event gotten from 'input') is, rather than
the new state, the state transition (i.e., as a key to select the right
MIDI_Event descendant). For example: when the transition is from program-
change to normal, the key to select the right MIDI_Event descendant is
PROGRAM_CHANGE() . '->' . NORMAL() [i.e., '2->0']. The MIDI_Event selected
will then be a ProgramChange_MIDI_Event, which will handle changing the
current event (which will be a note-off event) components such that the type
becomes program-change and the pitch becomes the patch number and then
outputting the event.

Began changes to implement this new design.  An incomplete summary of
the changes:

  - New class: MIDI_StateMachine, which is used (via Filter::Macro) to
	provide, essentially, (efficient) implementation in MIDI_EventFilter of
	the needed state-transition behavior (oddly, with a 'use
	MIDI_StateMachine' directive).
  - New class: BankSelect_MIDI_Event (implementation in-progress) to
	execute the needed behavior for the OVERRIDE -> BANK_SELECT (Note:
	BANK_SELECT is a new state.) transition.
  - New class: Null_MIDI_Event, to execute any transition that requires
	discarding the current (last-input) event - i.e., a no-op class.
  - Also used Filter::Macro in MIDI_Facilities so that the classes
	"inheriting from" MIDI_Facilities will have these facilities
	(constants, etc.) inline (i.e., expanded in the descendant class)
	for performance efficiency.  (I'm assuming efficiency will be an issue
	since this is essentially a real-time application.)
  - Changing to use, for constants, e.g., sub CONSTNAME { 2 } instead of
	the Readonly module, which is cleaner and perl optimizes (inlines)
	the use of empty prototypes that return a literal value.

The application compiles and runs, but, of course, does not run
correctly since it is not finished.