Most of the work of duplicating midifilter1 completed.
Basic functionality finished - it does pretty much what
../midi-alsa-work/midifilter1 does, although it has not yet been thoroughly
tested. The three descendants of MIDI_Event, BankSelect_MIDI_Event,
ProgramChange_MIDI_Event, and Static_MIDI_Event, have been implemented (at
least to the point of passing a basic test) and MIDI_EventFilter and
MIDI_StateMachine are finished (again, at least with respect to being able
to pass a basic test). (In particular, execute_state_change - in
MIDI_StateMachine - is complete in terms of, for the most part, matching the
functionality of the big MIDI-event inputting/routing while loop in
midifilter1.)

Some cleanup needs to be done, some bugs probably need to be discovered and
fixed, and, possibly, some implementation needs to be filled in; but this
version is basically functional and, if testing shows that this OO
design/implementation is likely to be sufficient in terms of both
flexibility/changeability and adequate real-time MIDI performance, this
project may suffice as a base for much added functionality in terms of MIDI
filtering, configurability, etc.  Performance seems adequate so far, but
much more testing is needed; also, there are probably some sections in the
code that can be changed/tuned for better performance if needed.