Provide way to cancel override mode - OVERRIDE -> NORMAL.
In override mode, if the current event is a note-off and its pitch is
not one that triggers a transition to program-change or to bank-select,
change state to normal.  This effectively cancels the override mode so
that, e.g., any keys the user strikes after that play as expected.  This
seems better (put another way, the user is likely to expect this) than
the previous behavior of staying in override mode until the user happens
to hit one of the program-change or bank-select keys/pitches.  (With the
previous behavior, if you hit the volume pedal [the event currently
programmed to trigger override state] and then change your mind - you
don't want to trigger a PC or BS after all - the only way to get back to
normal state is to hit a PC or BS key.  You can then (i.e., when in
override) effectively cancel by hitting a PC key and the key whose pitch
is for the current patch setting (so that effectively no PC actually
occurs), but that's awkward, especially if you're not sure which key to
hit for the current patch/program-number.)

In short, providing this effective "cancel override mode" facility
very likely provides a better and friendlier interface than the previous
behavior.