refactor execute_state_change, ExternalCommand_MIDI_Event, etc.
MIDI_StateMachine: Moved logic for the following state transitions out
of execute_state_change into their own functions (which return the new
state to execute_state_change):

	- override -> <whatever>
	- program-change -> <whatever>

For an original state of either override or program-change, no matter
what the new state will be, the response time of the system in executing the
state change and sending any resulting MIDI events to the connected output
ports does not need to be especially quick, since the user will not be
inputting real-time MIDI events during this time; rather, he will be
hitting the pedals-and/or-keys needed to invoke these state changes, one
at a time, and waiting until the whole "transaction" (such as a program
change) is complete before continuing to start hitting keys to perform
music again.  In other words, the real-time response is not nearly as
demanding as the states for which MIDI events need to be passed on right
away (i.e., the normal and bank-select states) - so these demanding
states remain handled in execute_state_change (to try to reduce latency
by avoiding a function call) while the override and program-change
states can be handled in a separate function.  The reason for this
change, of course, is to keep the size of the execute_state_change
function reasonable to reduce complexity and code maintenance time.
Tested successfully.

Added a new MIDI_Event descendant: ExternalCommand_MIDI_Event, plugged
it in to MIDI_EventFilter, and successfully tested it.  (It executes a
shell command.)  Added a new state to correspond to this command (which
is treated similarly to BankSelect_MIDI_Event) - EXTERNAL_CMD.

Added a few more constant definitions in MIDI_Facilities, such as START,
STOP, CONTINUE.

A few other small changes.