Added configurability, cleanup, etc.
Made many settings configurable, including the control-change control
number to trigger override mode (default: 7 - channel volume).  This
required a bit of refactoring/restructuring. FilterSpecification is a new
class that provides an interface with these settings. See 'config1' for an
example configuration file.  Removed the constants (and use of them) that
used to be used for (hard-coded) configuration.

Some cleanup, such as removing the old, commented-out switch-case blocks.
Removed Overriding_MIDI_Event - It's not needed since the result of changing
to the "override" state is a no-op.

Added a special <terminate> specification for an external command (handled
with ExternalCommand_MIDI_Event).  It results in termination of the
midifilter process.

Etc.