Better from/to port configuration
Enhancement to the from/to port configuration functionality to specify the
major port number (or whatever it's called) by name instead of number,
which allows the app to work correctly when the port number change (e.g.,
kmidimon is not always port 129).  The minor port number is still specified
as a number.  For example, to connect to rosegarden sub-port 0 (the "record
in" port):

to: rosegarden, 0

The old port specification style (e.g., 'to: 128, 0') is no longer valid -
config lines in that form will now be ignored.

config1 now uses this format.