filter-config-report command/config, etc.
Added new type of special "external" command (i.e., that triggers the
ExternalCommand_MIDI_Event): configured with the tag '<filter-config-report>'.
Invoking the MIDI message combination needed to trigger this command
results in a report of the current filtering-related triggers specified
in the config file.  The report is sent to the configured 'announcer'
(e.g., espeak).  (The purpose of this feature is to allow a user who has
not memorized his configuration to refresh his memory as to which MIDI
events [most of which will be note events with specific pitches] trigger
which filtering functionalities.)  The format in the config file to
specify which pitch (preceded, as usual, by the "override CC control-change"
event to trigger "override" mode) triggers this report is:

external_cmd: <pitch> <filter-config-report>

where <pitch> is a MIDI note-event pitch value and <filter-config-report> is
to be taken literally.  For example:

external_cmd: 88 <filter-config-report>

Also, changed the Announcer::announce routine to save the message in a
temporary file and then feed it to the 'announce_prog' (program
configured to make the "announcement") as the program's standard input.
This seems to be more reliable (at least, when using espeak) with
multi-line messages than passing the message to 'announce_prog' as an
argument.