Moved *pm files to a lib directory, etc.
Some restructuring: Moved the *.pm files to <dirname>/../lib/midifilter,
where <dirname> is the path of the midifilter script; and moved
midifilter to a ./bin directory.  Made the needed changes to @INC in
midifilter so that it will find the associated *.pm files (in
'<dirname>/../lib/midifilter').  With these changes, these files can be
installed as a package - e.g., where the main script is in
/usr/local/bin/midifilter and the *.pm files are in /usr/local/lib/midifilter;
or where midifilter is in $HOME/bin and the *pm files are in
$HOME/lib/midifilter.