Transposition logic bug fix
Bug fix: When two or more transposition specs overlap such that a particular
pitch is affected by all of these specs but only one of the specs has been
activated (i.e., the user has pressed a pedal/key combination activating
only that spec), one of the other overlapping specs was often applied,
incorrectly, during a transposition.  The problem was that the logic
identifying which spec was active (i.e., had been toggled "on" via
toggle_transposition) was missing - the notes affected by the transposition
were marked as "to be transposed" but the first transposition spec found
that contained the current pitch would be used during the actual
transposition instead of the correct one.  The fix was to change
toggle_transposition to associate each pitch within the range to be
transposed (instead of, as before, just marking the pitch with a boolean
status - to-be-transposed) with the pitch_value argument to
toggle_transposition (i.e., the value used as a key to id a specific
transposition spec [e.g., in: 'transpose_spec: 68 48..71 +7', the 68 is such
a key]) so that the key to the right transposition was available.  Then, in
'dispatch', retrieve that key (stored in the transposition_status array with
the current pitch as the index) and use it to retrieve the correct
transposition spec.