_set_state -> set_state; docs; config[12], etc.
Various small changes: In MIDI_StateMachine _set_state was changed to
set_state to indicate it is public, since it is called by other classes.
Made this change in the code that calls _set_state.  Removed some
debug/print statements in Regular_MIDI_Event that were inserted while
working on the transposition bug fix.  Filled in some more documentation
(README.md).  Tuned config1 a bit and added a new file, config2, which
sends to a2j_bridge, which on the JACK MIDI side is to be hooked up to
ardour - thus allowing midifilter to output to ardour (3).