Bug fix: negative patch# if bottom_note not set
If bottom_note is not configured, 36 is (apparently) used by default, which
causes a 'program_change_low' program change patch # (pitch) to be negative
if the key struck triggers a MIDI pitch less than 36. Added a check (in
ProgramChange_MIDI_Event.pm) for the resulting negative pitch/patch value
(set in MIDI_StateMachine.progchange_state_transition) to adjust a negative
pitch to 0 and output a warning message about it.