Improvement to program-change sample mode
A small change in behavior of "program-change sample mode":
When the sample mode is invoked, obtain the value of the last patch
that was explicitly set via a program-change event and set the initial
patch for sample mode to that value plus one, unless the last patch
value was 0, in which case, start at 0; if no program-change event has
occurred (i.e., there's no last patch value), start at 0.  This gives
the user the flexibility to start the sampling process wherever he
wants instead of always having to start at the first patch.